requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm>=4.64.0
tensorboard>=2.9.0
scikit-learn>=1.1.0
pandas>=1.4.0
seaborn>=0.11.0
opencv-python>=4.6.0
albumentations>=1.2.0
